+++ src/onegov.form/onegov/form/__init__.py
@@ -1,5 +1,10 @@
+import logging
+log = logging.getLogger('onegov.form')  # noqa
+log.addHandler(logging.NullHandler())  # noqa
+
 from onegov.form.collection import FormCollection
 from onegov.form.core import Form, with_options
+from onegov.form.display import render_field
 from onegov.form.models import (
     FormDefinition,
     FormSubmission,
@@ -14,5 +19,6 @@ __all__ = [
     'FormSubmission',
     'PendingFormSubmission',
     'CompleteFormSubmission',
+    'render_field',
     'with_options'
 ]